+++ packages/components-js/projects/components-wrapper/webpack.config.ts
@@ -20,6 +20,11 @@ const config: webpack.Configuration = {
     publicPath: `${deployUrl}/`,
   },
   optimization: {
+    // https://webpack.js.org/guides/code-splitting/
+    // https://webpack.js.org/plugins/split-chunks-plugin/
+    splitChunks: {
+      minSize: 30000,
+    },
     usedExports: true,
     chunkIds: false, // when this is set false, we need to provide a custom plugin to generate our chunkIds
     minimize: true,